+++ src/utils/logging/language-server-transport.ts
@@ -1,7 +1,6 @@
 import TransportStream from 'winston-transport';
 import { LEVEL, MESSAGE } from 'triple-beam';
 import type winston from 'winston';
-// eslint-disable-next-line n/no-missing-import
 import type { Connection, RemoteConsole } from 'vscode-languageserver/node';
 import type { TransportStreamOptions } from 'winston-transport';