+++ test/e2e/__tests__/format.ts
@@ -1,33 +1,37 @@
-import path from 'path';
+import * as assert from 'node:assert/strict';
+
 import { commands } from 'vscode';
-import { URI } from 'vscode-uri';
-import { ApiEvent } from '../../../src/extension/index';
+
+import { openDocument, closeAllEditors, setupSettings, sleep } from '../helpers';
 
 describe('Document formatting', () => {
-	// eslint-disable-next-line @typescript-eslint/no-var-requires
+	setupSettings({ '[css]': { 'editor.defaultFormatter': 'stylelint.vscode-stylelint' } });
+
+	afterEach(async () => {
+		await closeAllEditors();
+	});
+
+	// eslint-disable-next-line @typescript-eslint/no-require-imports
 	const isStylelint16 = require('stylelint/package.json').version.startsWith('16.');
 
 	// Only Stylelint before 16 had formatting related rules
 	const pre16It = isStylelint16 ? it.skip : it;
 
 	pre16It('should format document using formatting options', async () => {
-		const documentPath = path.resolve(workspaceDir, 'defaults/format.css');
-
-		const eventPromise = waitForApiEvent(
-			ApiEvent.DidRegisterDocumentFormattingEditProvider,
-			({ uri }) => path.relative(documentPath, URI.parse(uri).fsPath) === '',
-		);
-
-		const editor = await openDocument(documentPath);
-
-		await eventPromise;
+		const editor = await openDocument('defaults/format.css');
 
 		editor.options.tabSize = 4;
 		editor.options.insertSpaces = false;
 
+		await sleep(1000); // HACK: Prevent flaky test.
 		await commands.executeCommand('editor.action.formatDocument');
 
-		// eslint-disable-next-line jest/no-standalone-expect
-		expect(editor.document.getText()).toMatchSnapshot();
+		assert.equal(
+			editor.document.getText(),
+			`a {
+	color: #fff;
+}
+`,
+		);
 	});
 });